+++ AES.h
@@ -91,7 +91,7 @@ class AES {
   // data before calling, preferably using the padding methods listed below.
   // Decryption must use the same mode as the encryption.
   void Encrypt(const unsigned char* datain, unsigned char* dataout,
-               uint64_t numBlocks, BlockMode mode = CBC);
+               uint32_t numBlocks, BlockMode mode = CBC);
 
   // call this before any decryption with the key to use
   void StartDecryption(const unsigned char* key);
@@ -103,7 +103,7 @@ class AES {
   // must know the desired length of the output data, since all the blocks are
   // returned decrypted. Encryption must use the same mode as the decryption.
   void Decrypt(const unsigned char* datain, unsigned char* dataout,
-               uint64_t numBlocks, BlockMode mode = CBC);
+               uint32_t numBlocks, BlockMode mode = CBC);
 
  private:
   int Nb, Nk;  // block and key length / 32, should be 4,6,or 8
